requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.2.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
dj-rest-auth==7.0.0
Django==5.1.3
django-filter==24.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
flake8==7.1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
mccabe==0.7.0
packaging==24.2
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
pyOpenSSL==24.3.0
pytz==2024.2
PyYAML==6.0.2
service-identity==24.2.0
setuptools==75.6.0
sqlparse==0.5.2
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
zope.interface==7.2
